[not verified] set current time when changing autoplay start
